Isolate download and convert panels
