+++ gml_application_schema_toolbox/gui/convert_panel.py
@@ -23,13 +23,9 @@
 
 import os
 from osgeo import gdal
-import owslib
-from owslib.wfs import WebFeatureService
-from owslib.feature.wfs200 import WFSCapabilitiesReader
-from tempfile import NamedTemporaryFile
 
 from qgis.PyQt.QtCore import QSettings, Qt, QUrl, pyqtSlot, QFile, QIODevice, QAbstractItemModel, QModelIndex
-from qgis.PyQt.QtGui import QStandardItemModel, QStandardItem, QDesktopServices
+from qgis.PyQt.QtGui import QStandardItemModel, QStandardItem
 from qgis.PyQt.QtWidgets import QMessageBox, QFileDialog
 from qgis.PyQt.QtXml import QDomDocument
 from qgis.PyQt import uic
@@ -42,7 +38,7 @@ gmlasconf = os.path.join(os.path.dirname(__file__),
                          '..', 'conf', 'gmlasconf.xml')
 
 WIDGET, BASE = uic.loadUiType(os.path.join(
-    os.path.dirname(__file__), '..', 'ui', 'gmlas_dockwidget_base.ui'))
+    os.path.dirname(__file__), '..', 'ui', 'convert_panel.ui'))
 
 
 data_folder = '/home/qgis/qgisgmlas/data'
@@ -105,33 +101,20 @@ class PgsqlConnectionsModel(QAbstractItemModel):
         return self._groups()[index.row()]
 
 
-class GmlasPluginDockWidget(BASE, WIDGET):
+class ConvertPanel(BASE, WIDGET):
 
     def __init__(self, parent=None):
-        super(GmlasPluginDockWidget, self).__init__(parent)
+        super(ConvertPanel, self).__init__(parent)
         self.setupUi(self)
 
-        self.downloadProgressBar.setVisible(False)
         self.pgsqlFormWidget.setVisible(False)
         self.convertProgressBar.setVisible(False)
 
         self.pgsqlConnectionsBox.setModel(PgsqlConnectionsModel())
         self.pgsqlSchemaBox.setCurrentText('gmlas')
 
-        self.uriComboBox.addItem('http://geoserv.weichand.de:8080/geoserver/wfs')
-        self.uriComboBox.addItem('https://wfspoc.brgm-rec.fr/geoserver/ows')
-        self.uriComboBox.addItem('https://wfspoc.brgm-rec.fr/constellation/WS/wfs/BRGM:GWML2')
-        self.uriComboBox.addItem('http://geoserverref.brgm-rec.fr/geoserver/ows')
-        self.uriComboBox.addItem('http://minerals4eu.brgm-rec.fr/deegree/services/m4eu')
-
         self.gmlPathLineEdit.setText('/home/qgis/qgisgmlas/data/geosciml/mappedfeature.gml')
         self.sqlitePathLineEdit.setText('/home/qgis/qgisgmlas/data/test.sqlite')
-
-    def uri(self):
-        return self.uriComboBox.currentText()
-
-    def wfs(self):
-        return WebFeatureService(url=self.uri(), version='2.0.0')
 
     @pyqtSlot()
     def on_gmlPathButton_clicked(self):
@@ -147,59 +130,6 @@ class GmlasPluginDockWidget(BASE, WIDGET):
         XmlDialog(self, self.wfs().getcapabilities().read()).exec_()
         #url = WFSCapabilitiesReader().capabilities_url(self.uri())
         #QDesktopServices.openUrl(QUrl(url))
-
-    @pyqtSlot()
-    def on_downloadButton_clicked(self):
-        self.downloadProgressBar.setValue(0)
-        self.downloadProgressBar.setVisible(True)
-        self.setCursor(Qt.WaitCursor)
-        try:
-            out = self.download()
-        finally:
-            self.downloadProgressBar.setVisible(False)
-            self.unsetCursor()
-        if out is not None:
-            self.gmlPathLineEdit.setText(out)
-            self.tabWidget.setCurrentIndex(self.tabWidget.currentIndex() + 1)
-
-    def download(self):
-        wfs = self.wfs()
-
-        params = {}
-        #params['typename'] = ','.join(wfs.contents)
-        params['typename'] = sorted(wfs.contents.keys())[0]
-        if self.bboxGroupBox.isChecked():
-            params['bbox'] = '({})'.format(self.bboxLineEdit().text())
-        params['maxfeatures'] = self.featureLimitBox.value()
-        '''
-        srsname='urn:x-ogc:def:crs:EPSG:31468'
-        '''
-
-        try:
-            response = wfs.getfeature(**params)
-        except owslib.util.ServiceException as e:
-            QMessageBox.critical(self, 'ServiceException', str(e))
-            return
-        xml = response.read()
-
-        XmlDialog(self, xml).exec_()
-
-        doc = QDomDocument()
-        if not doc.setContent(xml):
-            return
-        root = doc.documentElement()
-        exception = root.firstChildElement('ows:Exception')
-        if not exception.isNull():
-            QMessageBox.critical(self,
-                                 'ows:Exception',
-                                 exception.text())
-            return
-
-        with NamedTemporaryFile(suffix='.gml', delete=False) as out:
-            out.write(bytes(xml, 'UTF-8'))
-            path = out.name
-
-        return path
 
     def gmlas_datasource_name(self):
         return "GMLAS:{}".format(self.gmlPathLineEdit.text())
